[UPDATE] Hashids

+++ config/hashids.php
@@ -5,7 +5,7 @@ return [
     'connections' => [
         'main'        => [
             'salt'     => env('HASHIDS_SALT', 'rexamnSuperSecretKey'),
-            'length'   => env('HASHIDS_LENGTH', 16),
+            'length'   => (int) env('HASHIDS_LENGTH', 16),
             'alphabet' => 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890',
         ],
         'alternative' => [
